backend/requirements.txt
#  How to install the dependencies
#  Write the command inside the cmd
#  Command : pip install -r requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
transformers==4.35.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
torch==2.0.1
google-generativeai==0.5.4 
joblib==1.3.2 
pandas==2.2.2